+++ src/sys/socket/addr.rs
@@ -13,7 +13,7 @@ use crate::sys::socket::addr::netlink::NetlinkAddr;
 use crate::sys::socket::addr::alg::AlgAddr;
 #[cfg(any(target_os = "ios", target_os = "macos"))]
 use std::os::unix::io::RawFd;
-#[cfg(any(target_os = "ios", target_os = "macos"))]
+#[cfg(all(feature = "ioctl", any(target_os = "ios", target_os = "macos")))]
 use crate::sys::socket::addr::sys_control::SysControlAddr;
 #[cfg(any(target_os = "android",
           target_os = "dragonfly",
@@ -25,6 +25,7 @@ use crate::sys::socket::addr::sys_control::SysControlAddr;
           target_os = "netbsd",
           target_os = "openbsd",
           target_os = "fuchsia"))]
+#[cfg(feature = "net")]
 pub use self::datalink::LinkAddr;
 #[cfg(any(target_os = "android", target_os = "linux"))]
 pub use self::vsock::VsockAddr;
@@ -262,6 +263,9 @@ impl AddressFamily {
         }
     }
 }
+
+feature!{
+#![feature = "net"]
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum InetAddr {
@@ -515,6 +519,7 @@ impl fmt::Display for Ipv6Addr {
         self.to_std().fmt(fmt)
     }
 }
+}
 
 /// A wrapper around `sockaddr_un`.
 #[derive(Clone, Copy, Debug)]
@@ -709,13 +714,16 @@ impl Hash for UnixAddr {
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 #[non_exhaustive]
 pub enum SockAddr {
+    #[cfg(feature = "net")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "net")))]
     Inet(InetAddr),
     Unix(UnixAddr),
     #[cfg(any(target_os = "android", target_os = "linux"))]
     Netlink(NetlinkAddr),
     #[cfg(any(target_os = "android", target_os = "linux"))]
     Alg(AlgAddr),
-    #[cfg(any(target_os = "ios", target_os = "macos"))]
+    #[cfg(all(feature = "ioctl", any(target_os = "ios", target_os = "macos")))]
+    #[cfg_attr(docsrs, doc(cfg(feature = "ioctl")))]
     SysControl(SysControlAddr),
     /// Datalink address (MAC)
     #[cfg(any(target_os = "android",
@@ -727,14 +735,19 @@ pub enum SockAddr {
               target_os = "illumos",
               target_os = "netbsd",
               target_os = "openbsd"))]
+    #[cfg(feature = "net")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "net")))]
     Link(LinkAddr),
     #[cfg(any(target_os = "android", target_os = "linux"))]
     Vsock(VsockAddr),
 }
 
 impl SockAddr {
+    feature!{
+    #![feature = "net"]
     pub fn new_inet(addr: InetAddr) -> SockAddr {
         SockAddr::Inet(addr)
+    }
     }
 
     pub fn new_unix<P: ?Sized + NixPath>(path: &P) -> Result<SockAddr> {
@@ -751,9 +764,12 @@ impl SockAddr {
         SockAddr::Alg(AlgAddr::new(alg_type, alg_name))
     }
 
+    feature!{
+    #![feature = "ioctl"]
     #[cfg(any(target_os = "ios", target_os = "macos"))]
     pub fn new_sys_control(sockfd: RawFd, name: &str, unit: u32) -> Result<SockAddr> {
         SysControlAddr::from_name(sockfd, name, unit).map(SockAddr::SysControl)
+    }
     }
 
     #[cfg(any(target_os = "android", target_os = "linux"))]
@@ -763,16 +779,20 @@ impl SockAddr {
 
     pub fn family(&self) -> AddressFamily {
         match *self {
+            #[cfg(feature = "net")]
             SockAddr::Inet(InetAddr::V4(..)) => AddressFamily::Inet,
+            #[cfg(feature = "net")]
             SockAddr::Inet(InetAddr::V6(..)) => AddressFamily::Inet6,
             SockAddr::Unix(..) => AddressFamily::Unix,
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Netlink(..) => AddressFamily::Netlink,
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Alg(..) => AddressFamily::Alg,
-            #[cfg(any(target_os = "ios", target_os = "macos"))]
+            #[cfg(all(feature = "ioctl",
+                      any(target_os = "ios", target_os = "macos")))]
             SockAddr::SysControl(..) => AddressFamily::System,
             #[cfg(any(target_os = "android", target_os = "linux"))]
+            #[cfg(feature = "net")]
             SockAddr::Link(..) => AddressFamily::Packet,
             #[cfg(any(target_os = "dragonfly",
                       target_os = "freebsd",
@@ -781,6 +801,7 @@ impl SockAddr {
                       target_os = "netbsd",
                       target_os = "illumos",
                       target_os = "openbsd"))]
+            #[cfg(feature = "net")]
             SockAddr::Link(..) => AddressFamily::Link,
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Vsock(..) => AddressFamily::Vsock,
@@ -802,23 +823,28 @@ impl SockAddr {
     /// unsafe because it takes a raw pointer as argument.  The caller must
     /// ensure that the pointer is valid.
     #[cfg(not(target_os = "fuchsia"))]
+    #[cfg(any(feature = "net", feature = "uio"))]
     pub(crate) unsafe fn from_libc_sockaddr(addr: *const libc::sockaddr) -> Option<SockAddr> {
         if addr.is_null() {
             None
         } else {
             match AddressFamily::from_i32(i32::from((*addr).sa_family)) {
                 Some(AddressFamily::Unix) => None,
+                #[cfg(feature = "net")]
                 Some(AddressFamily::Inet) => Some(SockAddr::Inet(
                     InetAddr::V4(*(addr as *const libc::sockaddr_in)))),
+                #[cfg(feature = "net")]
                 Some(AddressFamily::Inet6) => Some(SockAddr::Inet(
                     InetAddr::V6(*(addr as *const libc::sockaddr_in6)))),
                 #[cfg(any(target_os = "android", target_os = "linux"))]
                 Some(AddressFamily::Netlink) => Some(SockAddr::Netlink(
                     NetlinkAddr(*(addr as *const libc::sockaddr_nl)))),
-                #[cfg(any(target_os = "ios", target_os = "macos"))]
+                #[cfg(all(feature = "ioctl",
+                          any(target_os = "ios", target_os = "macos")))]
                 Some(AddressFamily::System) => Some(SockAddr::SysControl(
                     SysControlAddr(*(addr as *const libc::sockaddr_ctl)))),
                 #[cfg(any(target_os = "android", target_os = "linux"))]
+                #[cfg(feature = "net")]
                 Some(AddressFamily::Packet) => Some(SockAddr::Link(
                     LinkAddr(*(addr as *const libc::sockaddr_ll)))),
                 #[cfg(any(target_os = "dragonfly",
@@ -828,6 +854,7 @@ impl SockAddr {
                           target_os = "netbsd",
                           target_os = "illumos",
                           target_os = "openbsd"))]
+                #[cfg(feature = "net")]
                 Some(AddressFamily::Link) => {
                     let ether_addr = LinkAddr(*(addr as *const libc::sockaddr_dl));
                     if ether_addr.is_empty() {
@@ -855,6 +882,7 @@ impl SockAddr {
     /// a sockaddr * need to take the size of the underlying type as well and then internally cast it back.
     pub fn as_ffi_pair(&self) -> (&libc::sockaddr, libc::socklen_t) {
         match *self {
+            #[cfg(feature = "net")]
             SockAddr::Inet(InetAddr::V4(ref addr)) => (
                 // This cast is always allowed in C
                 unsafe {
@@ -862,6 +890,7 @@ impl SockAddr {
                 },
                 mem::size_of_val(addr) as libc::socklen_t
             ),
+            #[cfg(feature = "net")]
             SockAddr::Inet(InetAddr::V6(ref addr)) => (
                 // This cast is always allowed in C
                 unsafe {
@@ -892,7 +921,8 @@ impl SockAddr {
                 },
                 mem::size_of_val(sa) as libc::socklen_t
             ),
-            #[cfg(any(target_os = "ios", target_os = "macos"))]
+            #[cfg(all(feature = "ioctl",
+                      any(target_os = "ios", target_os = "macos")))]
             SockAddr::SysControl(SysControlAddr(ref sa)) => (
                 // This cast is always allowed in C
                 unsafe {
@@ -902,6 +932,7 @@ impl SockAddr {
 
             ),
             #[cfg(any(target_os = "android", target_os = "linux"))]
+            #[cfg(feature = "net")]
             SockAddr::Link(LinkAddr(ref addr)) => (
                 // This cast is always allowed in C
                 unsafe {
@@ -916,6 +947,7 @@ impl SockAddr {
                       target_os = "illumos",
                       target_os = "netbsd",
                       target_os = "openbsd"))]
+            #[cfg(feature = "net")]
             SockAddr::Link(LinkAddr(ref addr)) => (
                 // This cast is always allowed in C
                 unsafe {
@@ -938,13 +970,15 @@ impl SockAddr {
 impl fmt::Display for SockAddr {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match *self {
+            #[cfg(feature = "net")]
             SockAddr::Inet(ref inet) => inet.fmt(f),
             SockAddr::Unix(ref unix) => unix.fmt(f),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Netlink(ref nl) => nl.fmt(f),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Alg(ref nl) => nl.fmt(f),
-            #[cfg(any(target_os = "ios", target_os = "macos"))]
+            #[cfg(all(feature = "ioctl",
+                      any(target_os = "ios", target_os = "macos")))]
             SockAddr::SysControl(ref sc) => sc.fmt(f),
             #[cfg(any(target_os = "android",
                       target_os = "dragonfly",
@@ -955,6 +989,7 @@ impl fmt::Display for SockAddr {
                       target_os = "netbsd",
                       target_os = "illumos",
                       target_os = "openbsd"))]
+            #[cfg(feature = "net")]
             SockAddr::Link(ref ether_addr) => ether_addr.fmt(f),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Vsock(ref svm) => svm.fmt(f),
@@ -1060,6 +1095,8 @@ pub mod alg {
     }
 }
 
+feature!{
+#![feature = "ioctl"]
 #[cfg(any(target_os = "ios", target_os = "macos"))]
 pub mod sys_control {
     use crate::sys::socket::addr::AddressFamily;
@@ -1130,10 +1167,13 @@ pub mod sys_control {
         }
     }
 }
+}
 
 
 #[cfg(any(target_os = "android", target_os = "linux", target_os = "fuchsia"))]
 mod datalink {
+    feature!{
+    #![feature = "net"]
     use super::{fmt, AddressFamily};
 
     /// Hardware Address
@@ -1197,6 +1237,7 @@ mod datalink {
                 addr[5])
         }
     }
+    }
 }
 
 #[cfg(any(target_os = "dragonfly",
@@ -1207,6 +1248,8 @@ mod datalink {
           target_os = "netbsd",
           target_os = "openbsd"))]
 mod datalink {
+    feature!{
+    #![feature = "net"]
     use super::{fmt, AddressFamily};
 
     /// Hardware Address
@@ -1290,6 +1333,7 @@ mod datalink {
                 addr[4],
                 addr[5])
         }
+    }
     }
 }